Implement dynamic dispatch of operators
Each type has a "behavior", which is a function that obtains
property values: (value, propertyName) -> propertyValue.

Operators are implemented as property lookups.  Unary operators identify
properties: the property value is the result of unop.  Binary operators
identify methods: the property value is a function that accepts the second
binop argument.

Built-in types now support a reasonable set of operators.