Translate interpreter from Lua to JavaScript
I started with Lua partly because LPEG was a good starting point, but
then I ended up writing a custom PEG library (to deal with state).  Lua
is also a nice clean language (unlike JavaScript) and afforded better
build+test times, although those are not long enough to be of concern
right now (70 for Lua, 120ms with JS).

JavaScript makes sense now because it's about time to start working
through UI implications, so running the interpreter directly in the
browser or in Node (serving HTML to a browser) are both possibilities.