Fix TestJSB build class
When testing browser-targeted JS code off-target in Node, we load
mockdom.js prior to loading the target module.  Previously, we
tried to trigger this with a command-line expression, which we find
will fail when the target module does not import any other modules
(due to asynchronous loading of imports in JS, presumably).

The solution here is to have mockdom.js load the target module
after it has completed initialization.  Command lines get a bit
smaller, too.